backend/requirements.txt
Flask
tensorflow
numpy
opencv-python
scipy
scikit-image
scikit-learn
pymongo
python-dotenv